9689 zfs range lock code should not be zpl-specific
Reviewed by: Serapheim Dimitropoulos <serapheim.dimitro@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>
Approved by: Robert Mustacchi <rm@joyent.com>

The ZFS range locking code in zfs_rlock.c/h depends on ZPL-specific data
structures, specifically znode_t.  However, it's also used by the ZVOL
code, which uses a "dummy" znode_t to pass to the range locking code.

We should clean this up so that the range locking code is generic and
can be used equally by ZPL and ZVOL, and also can be used by future
consumers that may need to run in userland (libzpool) as well as the
kernel.

Closes #680