12241 smatch: misreported 'mask and shift to zero'
match_binop2() is testing expressions in form:
(mask & value) >> shift

and fails with variable >> shift, because we get incoming expression set
to: (mask & ) >> shift and the mask value may not be correct.

In specific error case we do have 64-bit lhs for >> operator, and we do
shift 32 bits; the mask value is set to 32-bit unsigned max and shifting it
by 32 bits, we do end up with match_binop2() reporting the warning.

Change-Id: I349ce56f0b82d45f0e016163332cfc2d3332e0b5