13710 create_ramdisk: prepare cpio boot archive for sparc
we can not gzip the cpio archive (should instead gzip files in it).
we need to add "bootblk" in front of cpio archive.

SPARC boot archive is treated as virtual disk, and it is assumed,
we do have boot block installed, which can read the file system
on that virtual disk.

Change-Id: I4ec2f306ba8974573f98a8466bc09af46aab72ed