stand: Use devformat rather than disk_devfmt
Fix layering violation and use devformat to get the string
representation of the device to see if we're mounted yet or not. Remove
added include to pickup disk.h.

Sponsored by:           Netflix
Reviewed by:            tsoome (prior version)
Differential Revision:  https://reviews.freebsd.org/D35919

Use devformat instead of disk_devfmt. This allows us to avoid knowing
the details of the device that's underneath us. Remove disk.h include
and the -I${LDRSRC} from the build of ufs.c since they are no longer
needed.

Sponsored by:           Netflix
Differential Revision:  https://reviews.freebsd.org/D35922

Change-Id: Ie6465fae2640c825c9d5648a6de4116bf5b6baeb