loader: udp receive should try harder to get packets for this connection
The current udp receive code does check if this packet does belong to us,
but it does return immediately if not - and this can trigger timeouts,
repeated send+receive calls and therefore delay or even fail the data
transfer.

Instead, we should read next packet and see if it does belong to us.

This patch should fix/improve network booting for systems, which are
receiving udp data streams during the network boot.

Change-Id: I3a3867a8eb179a88bb7d176d0663359a5b7f8e8d