rtld: don't store relocation types in anything less than a word
The audit missed instances of this outside libld, and then fatal
consequences were hidden by too many things being PIC.

AArch64 relocations are very sparse, and dynamic relocations have large
values.  Storing the relocation in a shorter type will truncate and we
will miss important clues (as well as utterly mis-processing them, but
that turns out to be surprisingly inconsequential!) about their nature.

Most importantly, our logic to avoid binding COPY relocations to
themselves fails because R_AARCH64_COPY has the value 1024, and so when
storred in a uchar_t takes a value of 0, R_AARCH64_NONE.  This means
that we never set LKUP_COPY when searching for the backing symbol for
these and bind them to themselves.  This would be bad enough, but
because the lists of COPY receiver and suppliers share memory in
Rt_map_copy (a given object may only be one, in correct operation), we
corrupt this utterly.