fix ld.so bootstrap and move .dynamic back to its normal place
Due to an annoyance of gas(1) the `_rt_boot` in ld.so was not correctly
implemented for the case where `.dynamic` precedes `.text` in
executables, as it traditionally does for us (but not, apparently, for
GNU ld).

The problem is that our calculation is based on calculating the offset
from our PC to `.dynamic` at compile time, and then applying that offset
to the runtime address.  Unfortunately, with gas(1), `.long` is just an
alias for `.int` and does not follow the ISA.  This means when the
offset was negative, we lost the high 32bits leaving it positive when we
came to do the addition.

The correct way to spell a 64bit value to gas(1) is `.quad`, use it and
make the comments clearer.