Use the regular vio* drivers on QEMU (FLAG DAY)
Previously, we'd been using custom `platmac` and `platmmc` drivers directly
linked to a modified set of virtio sources.  Split this out into a platform
`virtio` module (still with duplicated sources) earlier in the search path
than the regular `virtio`, and let the regular `vioblk` and `vioif` leaf
drivers do the work.

This will change your network device from `platmac<N>` to `vioif<N>` and
requires the `vioif` and `vioblk` drivers to be in the disk image.