merge the virtio implementation sources to ease future merges
This merges what used to be the `platvirtio` sources from Hayashi and the
regular common `virtio` sources by the simple expedient of splitting out the
pieces that are actually PCI v. MMIO dependent into separate source files
`virtio_pci.c` and `virtio_mmio.c` and specified which to use when building
the pair of drivers.

Hopefully, this will make future merges easier or at least provide a window on
what we should do to be able to support both more cleanly than this.