Fix thread specific data in nscd(8) (and possibly ldap_cachemgr(8))
`nscd` and `ldap_cachemgr` are both normally 32bit, and made assumptions that
`int` and a pointer are the same size, to store and retrieve their
thread-specific data.  Explicitly used `intptr_t` so that this is true
regardless of ISA.

I have no LDAP environment and have not tested `ldap_cachemgr`.