stand: Change disk_parsedev() API
Change the first argument to disk_parsedev() to be a pointer to a struct
devdesc *. This now gets filled in with a malloc'd structure that's
returned to the caller that the caller is repsonsible for freeing. Most
places in the tree passed in a malloc'd pointer anyway, and this moves
knowledge of disk_devdesc more firmly into the disk.[ch] code.

Sponsored by:           Netflix
Differential Revision:  https://reviews.freebsd.org/D37335

Change-Id: I5b85ca1feee2fe38279d8405a2f59aca89b93899