dis: Fix 'adrp' calculation
adrp was being handled in the same way as adr, that is the target value was
being calculated as the current address plus the immediate.
For adrp, the target is the page in which that resides.

Printing a label against adr/adrp was also not working, resulting in just a
stray '>' being printed.