dis: Print label for adrl pseudo, consolidate label printing code
This is a start of an attempt to show a useful label against adrl
(adrp + add), rather than showing a non-useful one against the 'adrp'.

For example, the existing code prints:

    get_crc_table:     60 00 00 90  adrp x0, 0xf000     <gzfwrite+0x20>
    get_crc_table+0x4: 00 60 1e 91  add x0, x0, #0x798

and the new code shows:

    get_crc_table:     60 00 00 90  adrp x0, 0xf000
    get_crc_table+0x4: 00 60 1e 91  add x0, x0, #0x798  <crc_table>

This is not complete and it doesn't handle all cases, but it's better than
before because 'gzfwrite' has nothing at all to do with the adrp.