dis: Split out last two parts of a64_dis_ldstr()
Looking at the function which is still not very readable after the last round
of cstyle, it appears that half of it has been split out to separate functions
but the second half has not. This appears to be work in progress by the
original author. Finish this and restore reasonable indentation.