Implement `panic_stopcpus` and `panic_idle`
This is the code that, when we crash, forces CPUs other than that on which we
crashed to spin idling.  While here, add a little bit of commentary to the
xcall code regarding some things that confuse me.

Note also that this, unlike other platforms, does not drop to CLOCK_LEVEL in
the idle loop, as that allows us to then take nested, lower priority,
interrupts on the quiesced CPU.

A comment has been added in `panic_idle` trying to explain this and noting
what we do instead and why.  Hopefully this becomes easier to reason about
once we have an actual debugger.