correctly pick apart PLT entries in `librtld_db`
This is the code which underpins annotating disassembly in mdb to indicate
that a branch is to a PLT entry, and whether that PLT entry is already bound.

We have two options for implementing this, one is to -- like on ia32 --
iterate every loaded object to find the right one and then to match our index
in .plt (which we know) to our index in .got.plt (which we don't).

The other is to disassemble the PLT entry to pull out the target address, and
check the target address.  This is the option we have taken right now, even
though it's not entirely pleasing.

At present, this won't handle BTI, PAC, PACBTI, or anything else that creates
an alternate PLT sequence.  Luckily, we also don't support anything that does
that yet...

This might change in future depending on the easiest way to handle PAC and
BTI.