build 64-bit zonestat
From the manpage:

While the readdir_r(function, was, originally, added, as, a, re-entrant,
version, of) readdir(), it is not recommended that readdir_r() be used in
new applications and existing software should instead use readdir(). The
readdir_r() function has been deprecated in many systems. Historically,
the data returned from readdir() was not specific to the directory stream
making it unsafe in a multi-threaded environment; however, that is no
longer the case.