elfedit: fix module search path
Fix the default `elfedit` search path so that we can find our modules.  This
changes the `%I` path which would previously expand to the 64bit ISA or the
empty string to only expand to the 64bit ISA on multilib systems.

This means it expands into the form we expect.

That is:
  ia32: %I -> ""
  amd64: %I -> "amd64"
  sparc: %I -> ""
  sparcv9: %I -> "sparcv9"
  aarch64: %I -> ""

Our handling of install paths for plugins on non-multilib systems is currently
a mess, and needs a longer term decision made and changes to match it.