fully relocate ld.so.1
For reasons I'm not certain of, `ld.so.1` only relocated `RELATIVE` and
`GLOB_DAT` (of which there should be done).  Most importantly, it didn't
relocate `JMP_SLOT`, which includes support for `LD_DEBUG` via `liblddbg.so`.

With this change we treat things like the other platforms do, and relocate
every relocation, with the understanding these should only be `RELATIVE`,
`JMP_SLOT`, or things handled in the same way.

This also makes us relocate ourselves in a manner much more similar to other
platforms (amd64), and so simplifies `boot.s` somewhat.